requirements.txt
flask>=3.0.0
line-bot-sdk>=2.4.0,<3.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.90.0
openai>=1.0.0
requests>=2.31.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
pytz>=2023.3
